requirements.txt
asgiref==3.9.1
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
charset-normalizer==3.4.3
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.5
fire==0.7.1
fonttools==4.59.1
lxml==6.0.1
numpy==2.2.6
opencv-python-headless==4.12.0.88
pdf2docx==0.5.8
pdf2image==1.17.0
pillow==11.3.0
pycparser==2.22
PyMuPDF==1.26.4
python-docx==1.2.0
reportlab==4.4.3
sqlparse==0.5.3
termcolor==3.1.0
tinycss2==1.4.0
typing_extensions==4.15.0
webencodings==0.5.1
